+++ samplesheets/assayapps/pep_ms/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class PepMsConfig(AppConfig):
+    name = 'samplesheets.assayapps.pep_ms'